+++ purch/api/routers/user.py
@@ -3,7 +3,7 @@ import uuid
 from typing import Annotated
 from fastapi import APIRouter, Response, Depends, HTTPException, status
 
-from purch.user.repository import UserRepository
+from purch.domains.user.repository import UserRepository
 from purch.domains.user.models import User
 from purch.auth.security import oauth2_scheme, get_current_active_user
 from purch.common.config import get_settings, Settings
