+++ purch/finance/tasks.py
@@ -3,10 +3,13 @@ from typing import Annotated
 from plaid.models import ItemGetRequest, AccountsGetRequest, TransactionsSyncRequest
 
 from purch.taskiq import broker
-from purch.core.models import User, Item, Account, Transaction
-from purch.finance.plaid import plaid_client
+from purch.domains.user.models import User
+from purch.domains.item.models import Item
+from purch.domains.account.models import Account
+from purch.domains.transaction.models import Transaction
+from purch.infrastructure.plaid.client import plaid_client
 from purch.finance.repository import FinanceRepository
-from purch.user.repository import UserRepository
+from purch.domains.user.repository import UserRepository
 from purch.common.config import get_settings
 from purch.common.logger import get_logger
